refactor(path): replace old fare with new fare package
