--- src/main/java/wooteco/subway/domain/fare/BasicFare.java
@@ -1,11 +1,0 @@
-package wooteco.subway.domain.fare;
-
-public final class BasicFare implements Fare {
-
-    private static final int VALUE = 1250;
-
-    @Override
-    public int calculate() {
-        return VALUE;
-    }
-}
