+++ src/main/java/wooteco/subway/domain/path/Path.java
@@ -3,7 +3,7 @@ package wooteco.subway.domain.path;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-import wooteco.subway.domain.fare2.Fare;
+import wooteco.subway.domain.fare.Fare;
 import wooteco.subway.domain.line.LineExtraFare;
 import wooteco.subway.domain.section.Section;
 import wooteco.subway.domain.station.Station;
